Reorganized the repository and got php working on my code.
I organized the repository. I managed to get php working in my code for the website. I put instructions in the txt file on how to get the website working for you. The up to date code is in the To-DoListWebsite folder. Just put that folder in your htdocs folder.